chore: reorganise ai agent edit page & evals
