+++ packages/frontend/src/ee/pages/AiAgents/ProjectAiAgentEditPage.tsx
@@ -1,78 +1,44 @@
-import { FeatureFlags } from '@lightdash/common';
+import { ProjectType } from '@lightdash/common';
 import {
-    ActionIcon,
-    Anchor,
-    Badge,
+    AppShell,
     Box,
     Button,
-    Card,
-    Code,
-    Collapse,
     Container,
     Group,
-    HoverCard,
-    Loader,
-    LoadingOverlay,
-    MultiSelect,
     Paper,
     Stack,
-    Switch,
-    Tabs,
-    TagsInput,
     Text,
-    Textarea,
-    TextInput,
     Title,
-    Tooltip,
 } from '@mantine-8/core';
-import { useDisclosure } from '@mantine-8/hooks';
 import { useForm, zodResolver } from '@mantine/form';
 import {
     IconAdjustmentsAlt,
-    IconAlertTriangle,
     IconArrowLeft,
     IconBook2,
-    IconCheck,
-    IconExternalLink,
-    IconInfoCircle,
-    IconLock,
-    IconPlug,
-    IconPointFilled,
-    IconRefresh,
-    IconTrash,
+    IconMessageCircleShare,
 } from '@tabler/icons-react';
-import { useCallback, useEffect, useMemo, useState, type FC } from 'react';
-import { useNavigate, useParams } from 'react-router';
+import { useEffect, type FC } from 'react';
+import { Link, useLocation, useNavigate, useParams } from 'react-router';
 import { z } from 'zod';
 import { LightdashUserAvatar } from '../../../components/Avatar';
-import LinkButton from '../../../components/common/LinkButton';
 import MantineIcon from '../../../components/common/MantineIcon';
-import MantineModal from '../../../components/common/MantineModal';
-import { useGetSlack, useSlackChannels } from '../../../hooks/slack/useSlack';
-import { useFeatureFlag } from '../../../hooks/useFeatureFlagEnabled';
-import { useOrganizationGroups } from '../../../hooks/useOrganizationGroups';
-import { useProject } from '../../../hooks/useProject';
+import {
+    BANNER_HEIGHT,
+    NAVBAR_HEIGHT,
+} from '../../../components/common/Page/constants';
+import { useProjects } from '../../../hooks/useProjects';
 import useApp from '../../../providers/App/useApp';
-import { UserAccessMultiSelect } from '../../components/UserAccessMultiSelect';
-import { AiAgentEditPageLayout } from '../../features/aiCopilot/components/AiAgentEditPageLayout/AiAgentEditPageLayout';
+import { AiAgentFormSetup } from '../../features/aiCopilot/components/AiAgentFormSetup';
 import { EvalDetail } from '../../features/aiCopilot/components/Evals/EvalDetail';
 import { EvalRunDetails } from '../../features/aiCopilot/components/Evals/EvalRunDetails';
-import { EvalTabLayout } from '../../features/aiCopilot/components/Evals/EvalTabLayout';
-import {
-    InstructionsGuidelines,
-    InstructionsTemplates,
-} from '../../features/aiCopilot/components/InstructionsSupport';
+import { EvalSectionLayout } from '../../features/aiCopilot/components/Evals/EvalSectionLayout';
 import { useAiAgentPermission } from '../../features/aiCopilot/hooks/useAiAgentPermission';
 import {
-    useDeleteAiAgentMutation,
     useProjectAiAgent,
-    useProjectAiAgents,
     useProjectCreateAiAgentMutation,
     useProjectUpdateAiAgentMutation,
 } from '../../features/aiCopilot/hooks/useProjectAiAgents';
-import { useGetAgentExploreAccessSummary } from '../../features/aiCopilot/hooks/useUserAgentPreferences';
-import AiExploreAccessTree from './AiExploreAccessTree';
-import { EvalsTab } from './EvalsTab';
+import { EvalsSetup } from './EvalsSetup';
 
 const formSchema = z.object({
     name: z.string().min(1),
@@ -95,97 +61,30 @@ type Props = {
 };
 
 const ProjectAiAgentEditPage: FC<Props> = ({ isCreateMode = false }) => {
+    const navigate = useNavigate();
     const { agentUuid, projectUuid, evalUuid, runUuid } = useParams<{
         agentUuid: string;
         projectUuid: string;
         evalUuid?: string;
         runUuid?: string;
     }>();
+    const location = useLocation();
     const canManageAgents = useAiAgentPermission({
         action: 'manage',
         projectUuid,
     });
-
-    const navigate = useNavigate();
+    const { data: projects } = useProjects();
+    const isCurrentProjectPreview = !!projects?.find(
+        (project) =>
+            project.projectUuid === projectUuid &&
+            project.type === ProjectType.PREVIEW,
+    );
     const { user } = useApp();
 
-    const [deleteModalOpen, setDeleteModalOpen] = useState(false);
-
-    const { data: project } = useProject(projectUuid);
-    const { mutateAsync: createAgent, isLoading: isCreating } =
-        useProjectCreateAiAgentMutation(projectUuid!);
-    const { mutateAsync: updateAgent, isLoading: isUpdating } =
-        useProjectUpdateAiAgentMutation(projectUuid!);
-    const { mutateAsync: deleteAgent } = useDeleteAiAgentMutation(projectUuid!);
-
     const actualAgentUuid = !isCreateMode && agentUuid ? agentUuid : undefined;
-
     const { data: agent, isLoading: isLoadingAgent } = useProjectAiAgent(
         projectUuid,
         actualAgentUuid,
-    );
-
-    const { data: slackInstallation, isLoading: isLoadingSlackInstallation } =
-        useGetSlack();
-
-    const { data: agents, isSuccess: isSuccessAgents } = useProjectAiAgents({
-        projectUuid,
-        redirectOnUnauthorized: true,
-    });
-
-    const userGroupsFeatureFlagQuery = useFeatureFlag(
-        FeatureFlags.UserGroupsEnabled,
-    );
-
-    const isGroupsEnabled =
-        userGroupsFeatureFlagQuery.isSuccess &&
-        userGroupsFeatureFlagQuery.data.enabled;
-
-    const { data: groups, isLoading: isLoadingGroups } = useOrganizationGroups(
-        {
-            includeMembers: 5,
-        },
-        {
-            enabled: isGroupsEnabled,
-        },
-    );
-
-    const {
-        data: slackChannels,
-        refresh: refreshChannels,
-        isRefreshing,
-        isLoading: isLoadingSlackChannels,
-    } = useSlackChannels(
-        '',
-        {
-            excludeArchived: true,
-            excludeDms: true,
-            excludeGroups: true,
-        },
-        {
-            enabled: !!slackInstallation?.organizationUuid && isSuccessAgents,
-        },
-    );
-
-    const slackChannelOptions = useMemo(
-        () =>
-            slackChannels?.map((channel) => ({
-                value: channel.id,
-                label: channel.name,
-                disabled: agents?.some((a) =>
-                    a.integrations.some((i) => i.channelId === channel.id),
-                ),
-            })) ?? [],
-        [slackChannels, agents],
-    );
-
-    const groupOptions = useMemo(
-        () =>
-            groups?.map((group) => ({
-                value: group.uuid,
-                label: group.name,
-            })) ?? [],
-        [groups],
     );
 
     const form = useForm<z.infer<typeof formSchema>>({
@@ -224,28 +123,15 @@ const ProjectAiAgentEditPage: FC<Props> = ({ isCreateMode = false }) => {
         // eslint-disable-next-line react-hooks/exhaustive-deps
     }, [agent, isCreateMode]);
 
-    const [isExploreAccessSummaryOpen, { toggle: toggleExploreAccessSummary }] =
-        useDisclosure(false);
+    // Derive activeTab from current pathname
+    const activeTab = location.pathname.includes('/evals') ? 'evals' : 'setup';
 
-    const exploreAccessSummaryQuery = useGetAgentExploreAccessSummary(
+    const { mutateAsync: createAgent } = useProjectCreateAiAgentMutation(
         projectUuid!,
-        {
-            tags: form.values.tags,
-        },
     );
-
-    const slackChannelsConfigured = useMemo(
-        () =>
-            form.values.integrations.some(
-                (i) => i.type === 'slack' && i.channelId,
-            ),
-        [form.values.integrations],
+    const { mutateAsync: updateAgent } = useProjectUpdateAiAgentMutation(
+        projectUuid!,
     );
-
-    const handleBack = () => {
-        void navigate(-1);
-    };
-
     const handleSubmit = form.onSubmit(async (values) => {
         if (!projectUuid || !user?.data) {
             return;
@@ -256,7 +142,9 @@ const ProjectAiAgentEditPage: FC<Props> = ({ isCreateMode = false }) => {
                 ...values,
                 projectUuid,
             });
-        } else if (actualAgentUuid) {
+        }
+
+        if (actualAgentUuid) {
             await updateAgent({
                 uuid: actualAgentUuid,
                 projectUuid,
@@ -264,24 +152,6 @@ const ProjectAiAgentEditPage: FC<Props> = ({ isCreateMode = false }) => {
             });
         }
     });
-
-    const handleDeleteClick = useCallback(() => {
-        setDeleteModalOpen(true);
-    }, []);
-
-    const handleDelete = useCallback(async () => {
-        if (!actualAgentUuid || !user?.data || !projectUuid || !agent) {
-            return;
-        }
-
-        await deleteAgent(actualAgentUuid);
-
-        setDeleteModalOpen(false);
-    }, [actualAgentUuid, deleteAgent, user?.data, projectUuid, agent]);
-
-    const handleCancelDelete = useCallback(() => {
-        setDeleteModalOpen(false);
-    }, []);
 
     useEffect(() => {
         if (!canManageAgents) {
@@ -297,7 +167,9 @@ const ProjectAiAgentEditPage: FC<Props> = ({ isCreateMode = false }) => {
                         <Button
                             variant="subtle"
                             leftSection={<MantineIcon icon={IconArrowLeft} />}
-                            onClick={handleBack}
+                            onClick={() =>
+                                navigate(`/projects/${projectUuid}/ai-agents`)
+                            }
                         >
                             Back to Agents
                         </Button>
@@ -322,766 +194,141 @@ const ProjectAiAgentEditPage: FC<Props> = ({ isCreateMode = false }) => {
     }
 
     return (
-        <AiAgentEditPageLayout
-            header={
-                <Stack gap="xs" align="flex-start">
-                    <Button
-                        variant="subtle"
-                        leftSection={<MantineIcon icon={IconArrowLeft} />}
-                        onClick={handleBack}
-                    >
-                        Back
-                    </Button>
-
-                    <Group
-                        justify="space-between"
-                        wrap="nowrap"
-                        align="center"
-                        w="100%"
-                    >
-                        <Group gap="sm" align="center" flex="1" wrap="nowrap">
-                            <LightdashUserAvatar
-                                name={isCreateMode ? '+' : form.values.name}
-                                variant="filled"
-                                src={
-                                    !isCreateMode
-                                        ? form.values.imageUrl
-                                        : undefined
-                                }
-                                size={48}
-                            />
-                            <Stack gap={0}>
-                                <Group gap="xs" wrap="nowrap">
-                                    <Title order={2} lineClamp={1} w="100%">
-                                        {isCreateMode
-                                            ? 'New Agent'
-                                            : agent?.name || 'Agent'}
-                                    </Title>
-
-                                    {!isCreateMode && agent && (
-                                        <LinkButton
-                                            href={`/ai-agents/admin?agents=${agent?.uuid}&projects=${projectUuid}`}
-                                            target="_blank"
-                                            variant="light"
-                                            size="sm"
-                                            leftIcon={IconExternalLink}
-                                        >
-                                            Conversations
-                                        </LinkButton>
-                                    )}
-                                </Group>
-                                <Text size="xs" c="dimmed">
-                                    Last modified:{' '}
-                                    {new Date(
-                                        agent?.updatedAt ?? new Date(),
-                                    ).toLocaleString()}
-                                </Text>
-                            </Stack>
-                        </Group>
-                        <Group justify="flex-end" gap="xs">
-                            <Button
-                                variant="outline"
-                                onClick={() => handleSubmit()}
-                                loading={isCreating || isUpdating}
-                                leftSection={<MantineIcon icon={IconCheck} />}
-                                disabled={
-                                    isCreateMode
-                                        ? !form.isValid()
-                                        : !form.isDirty()
-                                }
-                            >
-                                {isCreateMode ? 'Create agent' : 'Save changes'}
-                            </Button>
-                        </Group>
-                    </Group>
-                </Stack>
-            }
+        <AppShell
+            padding="md"
+            header={{
+                height:
+                    NAVBAR_HEIGHT +
+                    (isCurrentProjectPreview ? BANNER_HEIGHT : 0),
+            }}
+            navbar={{
+                width: 300,
+                breakpoint: 'sm',
+            }}
         >
-            <Stack gap="xs">
-                <Tabs
-                    defaultValue={evalUuid || runUuid ? 'evals' : 'setup'}
-                    keepMounted={false}
-                >
-                    {!isCreateMode && (
-                        <Tabs.List>
-                            <Tabs.Tab value="setup">Setup</Tabs.Tab>
-                            <Tabs.Tab value="evals">Evals</Tabs.Tab>
-                        </Tabs.List>
+            <AppShell.Navbar p="md">
+                <Stack gap="lg">
+                    <Stack gap="md" align="center">
+                        <Group justify="space-between" w="100%">
+                            <Button
+                                variant="subtle"
+                                color="gray.4"
+                                size="xs"
+                                leftSection={
+                                    <MantineIcon icon={IconArrowLeft} />
+                                }
+                                onClick={() => {
+                                    if (isCreateMode) {
+                                        void navigate(
+                                            `/projects/${projectUuid}/ai-agents`,
+                                        );
+                                    } else {
+                                        void navigate(
+                                            `/projects/${projectUuid}/ai-agents/${actualAgentUuid}`,
+                                        );
+                                    }
+                                }}
+                            >
+                                Back
+                            </Button>
+                            {form.isDirty() && (
+                                <Button
+                                    size="xs"
+                                    disabled={!form.isDirty()}
+                                    onClick={() => handleSubmit()}
+                                >
+                                    Save changes
+                                </Button>
+                            )}
+                        </Group>
+                        <LightdashUserAvatar
+                            name={isCreateMode ? '+' : form.values.name}
+                            variant="filled"
+                            src={
+                                !isCreateMode ? form.values.imageUrl : undefined
+                            }
+                            size={80}
+                        />
+                        <Stack gap="xs" align="center">
+                            <Title order={4} ta="center" lineClamp={2}>
+                                {isCreateMode
+                                    ? 'New Agent'
+                                    : agent?.name || 'Agent'}
+                            </Title>
+                        </Stack>
+                    </Stack>
+
+                    <Stack gap="xs">
+                        <Button
+                            variant={activeTab === 'setup' ? 'light' : 'subtle'}
+                            fullWidth
+                            justify="flex-start"
+                            leftSection={
+                                <MantineIcon icon={IconAdjustmentsAlt} />
+                            }
+                            onClick={() => {
+                                if (isCreateMode) {
+                                    // In create mode, we can't navigate to agent-specific routes
+                                    return;
+                                }
+                                void navigate(
+                                    `/projects/${projectUuid}/ai-agents/${actualAgentUuid}/edit`,
+                                );
+                            }}
+                        >
+                            Setup
+                        </Button>
+                        {!isCreateMode && (
+                            <Button
+                                variant={
+                                    activeTab === 'evals' ? 'light' : 'subtle'
+                                }
+                                fullWidth
+                                justify="flex-start"
+                                leftSection={<MantineIcon icon={IconBook2} />}
+                                onClick={() => {
+                                    void navigate(
+                                        `/projects/${projectUuid}/ai-agents/${actualAgentUuid}/edit/evals`,
+                                    );
+                                }}
+                            >
+                                Evals
+                            </Button>
+                        )}
+                        {!isCreateMode && (
+                            <Button
+                                fullWidth
+                                variant="subtle"
+                                justify="flex-start"
+                                leftSection={
+                                    <MantineIcon
+                                        icon={IconMessageCircleShare}
+                                    />
+                                }
+                                component={Link}
+                                to={`/ai-agents/admin?projects=${projectUuid}&agents=${actualAgentUuid}`}
+                            >
+                                Conversations
+                            </Button>
+                        )}
+                    </Stack>
+                </Stack>
+            </AppShell.Navbar>
+            <AppShell.Main pt={0} pr={0}>
+                <Box>
+                    {activeTab === 'setup' && (
+                        <Box pt="sm" pr="sm">
+                            <AiAgentFormSetup
+                                mode={isCreateMode ? 'create' : 'edit'}
+                                agentUuid={actualAgentUuid!}
+                                form={form}
+                                projectUuid={projectUuid!}
+                            />
+                        </Box>
                     )}
 
-                    <Tabs.Panel value="setup" pt="lg">
-                        <form>
-                            <Stack gap="sm">
-                                <Paper p="xl">
-                                    <Group align="center" gap="xs" mb="md">
-                                        <Paper p="xxs" withBorder radius="sm">
-                                            <MantineIcon
-                                                icon={IconAdjustmentsAlt}
-                                                size="md"
-                                            />
-                                        </Paper>
-                                        <Title order={5} c="gray.9" fw={700}>
-                                            Basic information
-                                        </Title>
-                                    </Group>
-                                    <Stack>
-                                        <Group>
-                                            <TextInput
-                                                label="Agent Name"
-                                                placeholder="Enter a name for this agent"
-                                                {...form.getInputProps('name')}
-                                                style={{ flexGrow: 1 }}
-                                                variant="subtle"
-                                            />
-                                            <Tooltip label="Agents can only be created in the context the current project">
-                                                <TextInput
-                                                    label="Project"
-                                                    placeholder="Enter a project"
-                                                    value={project?.name}
-                                                    readOnly
-                                                    style={{ flexGrow: 1 }}
-                                                    variant="subtle"
-                                                />
-                                            </Tooltip>
-                                        </Group>
-                                        <TextInput
-                                            style={{ flexGrow: 1 }}
-                                            miw={200}
-                                            variant="subtle"
-                                            label="Avatar image URL"
-                                            description="Please provide an image url like https://example.com/avatar.jpg. If not provided, a default avatar will be used."
-                                            placeholder="https://example.com/avatar.jpg"
-                                            type="url"
-                                            {...form.getInputProps('imageUrl')}
-                                            onChange={(e) => {
-                                                const value = e.target.value;
-                                                form.setFieldValue(
-                                                    'imageUrl',
-                                                    value ? value : null,
-                                                );
-                                            }}
-                                        />
-                                    </Stack>
-                                </Paper>
-
-                                <Paper p="xl">
-                                    <Group align="center" gap="xs" mb="md">
-                                        <Paper p="xxs" withBorder radius="sm">
-                                            <MantineIcon
-                                                icon={IconLock}
-                                                size="md"
-                                            />
-                                        </Paper>
-                                        <Title order={5} c="gray.9" fw={700}>
-                                            Access control
-                                        </Title>
-                                    </Group>
-                                    <Stack>
-                                        <UserAccessMultiSelect
-                                            projectUuid={projectUuid!}
-                                            isGroupsEnabled={isGroupsEnabled}
-                                            value={form.values.userAccess}
-                                            onChange={(value) => {
-                                                form.setFieldValue(
-                                                    'userAccess',
-                                                    value,
-                                                );
-                                            }}
-                                        />
-
-                                        {isGroupsEnabled && (
-                                            <Stack gap="xs">
-                                                <MultiSelect
-                                                    variant="subtle"
-                                                    label={
-                                                        <Group gap="xs">
-                                                            <Text
-                                                                fz="sm"
-                                                                fw={500}
-                                                            >
-                                                                Group Access
-                                                            </Text>
-                                                            <Tooltip
-                                                                label="Admins and developers will always have access to this agent."
-                                                                withArrow
-                                                                withinPortal
-                                                                multiline
-                                                                position="right"
-                                                                maw="250px"
-                                                            >
-                                                                <MantineIcon
-                                                                    icon={
-                                                                        IconInfoCircle
-                                                                    }
-                                                                />
-                                                            </Tooltip>
-                                                        </Group>
-                                                    }
-                                                    description="Select groups that can access this agent."
-                                                    placeholder={
-                                                        isLoadingGroups
-                                                            ? 'Loading groups...'
-                                                            : groupOptions.length ===
-                                                              0
-                                                            ? 'No groups available'
-                                                            : 'Select groups or leave empty for all users'
-                                                    }
-                                                    data={groupOptions}
-                                                    disabled={
-                                                        isLoadingGroups ||
-                                                        groupOptions.length ===
-                                                            0
-                                                    }
-                                                    comboboxProps={{
-                                                        transitionProps: {
-                                                            transition: 'pop',
-                                                            duration: 200,
-                                                        },
-                                                    }}
-                                                    clearable
-                                                    {...form.getInputProps(
-                                                        'groupAccess',
-                                                    )}
-                                                    value={
-                                                        form.getInputProps(
-                                                            'groupAccess',
-                                                        ).value ?? []
-                                                    }
-                                                    onChange={(value) => {
-                                                        form.setFieldValue(
-                                                            'groupAccess',
-                                                            value.length > 0
-                                                                ? value
-                                                                : [],
-                                                        );
-                                                    }}
-                                                />
-                                            </Stack>
-                                        )}
-
-                                        <Box>
-                                            <TagsInput
-                                                variant="subtle"
-                                                label={
-                                                    <Group gap="xs">
-                                                        <Text fz="sm" fw={500}>
-                                                            Tags
-                                                        </Text>
-                                                        <HoverCard
-                                                            position="right"
-                                                            withArrow
-                                                        >
-                                                            <HoverCard.Target>
-                                                                <MantineIcon
-                                                                    icon={
-                                                                        IconInfoCircle
-                                                                    }
-                                                                />
-                                                            </HoverCard.Target>
-                                                            <HoverCard.Dropdown maw="250px">
-                                                                <Text fz="xs">
-                                                                    Add tags to
-                                                                    control
-                                                                    which
-                                                                    metrics and
-                                                                    dimensions
-                                                                    your AI
-                                                                    agent can
-                                                                    access. See
-                                                                    more in our{' '}
-                                                                    <Anchor
-                                                                        fz="xs"
-                                                                        c="dimmed"
-                                                                        underline="always"
-                                                                        href="https://docs.lightdash.com/guides/ai-agents#limiting-access-to-specific-explores-and-fields"
-                                                                        target="_blank"
-                                                                    >
-                                                                        docs
-                                                                    </Anchor>
-                                                                </Text>
-                                                            </HoverCard.Dropdown>
-                                                        </HoverCard>
-                                                    </Group>
-                                                }
-                                                placeholder="Select tags"
-                                                inputWrapperOrder={[
-                                                    'label',
-                                                    'input',
-                                                    'description',
-                                                ]}
-                                                description={
-                                                    exploreAccessSummaryQuery.isSuccess ? (
-                                                        exploreAccessSummaryQuery
-                                                            .data.length ===
-                                                        0 ? (
-                                                            'No explorers are available for this tag selection. Make sure to use the correct tags, or tag the project with the correct tags and redeploy the project.'
-                                                        ) : (
-                                                            <>
-                                                                {
-                                                                    exploreAccessSummaryQuery
-                                                                        .data
-                                                                        .length
-                                                                }{' '}
-                                                                explores will be
-                                                                available to
-                                                                this agent.{' '}
-                                                                <Anchor
-                                                                    size="xs"
-                                                                    onClick={
-                                                                        toggleExploreAccessSummary
-                                                                    }
-                                                                >
-                                                                    Click here
-                                                                </Anchor>{' '}
-                                                                to see detailed
-                                                                list with
-                                                                metrics and
-                                                                dimensions.
-                                                            </>
-                                                        )
-                                                    ) : (
-                                                        `Loading AI access information...`
-                                                    )
-                                                }
-                                                {...form.getInputProps('tags')}
-                                                value={
-                                                    form.getInputProps('tags')
-                                                        .value ?? []
-                                                }
-                                                onChange={(value) => {
-                                                    form.setFieldValue(
-                                                        'tags',
-                                                        value.length > 0
-                                                            ? value
-                                                            : null,
-                                                    );
-                                                }}
-                                            />
-
-                                            {exploreAccessSummaryQuery.isSuccess ? (
-                                                <Collapse
-                                                    mt="xs"
-                                                    in={
-                                                        isExploreAccessSummaryOpen
-                                                    }
-                                                >
-                                                    <Card>
-                                                        <AiExploreAccessTree
-                                                            exploreAccessSummary={
-                                                                exploreAccessSummaryQuery.data
-                                                            }
-                                                        />
-                                                    </Card>
-                                                </Collapse>
-                                            ) : null}
-                                        </Box>
-
-                                        <Stack gap="sm">
-                                            <Box>
-                                                <Group gap="xs" c="gray.7">
-                                                    <Title
-                                                        order={6}
-                                                        size="sm"
-                                                        fw={500}
-                                                    >
-                                                        Enable Data Access
-                                                    </Title>
-                                                    <Tooltip
-                                                        label="When enabled, the AI agent can analyze chart data and provide insights. When disabled, the agent only creates visualizations without accessing the underlying data."
-                                                        withArrow
-                                                        withinPortal
-                                                        multiline
-                                                        position="right"
-                                                        maw="300px"
-                                                    >
-                                                        <MantineIcon
-                                                            icon={
-                                                                IconInfoCircle
-                                                            }
-                                                        />
-                                                    </Tooltip>
-                                                    <Badge
-                                                        size="xs"
-                                                        color="violet"
-                                                        variant="light"
-                                                    >
-                                                        New
-                                                    </Badge>
-                                                </Group>
-                                                <Text c="dimmed" size="xs">
-                                                    <>
-                                                        Allows the agent to
-                                                        access and analyze the
-                                                        actual data behind
-                                                        charts to provide
-                                                        detailed insights and
-                                                        answer questions about
-                                                        the data.{' '}
-                                                        <Anchor
-                                                            href="https://docs.lightdash.com/guides/ai-agents#data-access-control"
-                                                            target="_blank"
-                                                            size="xs"
-                                                        >
-                                                            Learn more
-                                                        </Anchor>
-                                                    </>
-                                                </Text>
-                                            </Box>
-                                            <Switch
-                                                variant="subtle"
-                                                color="blue.4"
-                                                label={
-                                                    form.values.enableDataAccess
-                                                        ? 'Enabled'
-                                                        : 'Disabled'
-                                                }
-                                                {...form.getInputProps(
-                                                    'enableDataAccess',
-                                                    {
-                                                        type: 'checkbox',
-                                                    },
-                                                )}
-                                            />
-                                        </Stack>
-                                    </Stack>
-                                </Paper>
-
-                                <Paper p="xl">
-                                    <Stack gap="md">
-                                        <Group align="center" gap="xs">
-                                            <Paper
-                                                p="xxs"
-                                                withBorder
-                                                radius="sm"
-                                            >
-                                                <MantineIcon
-                                                    icon={IconBook2}
-                                                    size="md"
-                                                />
-                                            </Paper>
-                                            <Title
-                                                order={5}
-                                                c="gray.9"
-                                                fw={700}
-                                            >
-                                                Knowledge & expertise
-                                            </Title>
-                                        </Group>
-                                        <Stack gap="xs">
-                                            <Textarea
-                                                variant="subtle"
-                                                label="Instructions"
-                                                description="Set the overall behavior and task for the agent. This defines how it should respond and what its purpose is."
-                                                placeholder="You are a marketing analytics expert. Focus on campaign performance, customer acquisition costs, and ROI metrics. Always use bar charts and tables to visualize data."
-                                                resize="vertical"
-                                                autosize
-                                                minRows={3}
-                                                maxRows={8}
-                                                {...form.getInputProps(
-                                                    'instruction',
-                                                )}
-                                            />
-                                            <Text size="xs" c="dimmed">
-                                                {form.values.instruction
-                                                    ?.length ?? 0}{' '}
-                                                characters
-                                            </Text>
-                                        </Stack>
-                                        <Stack gap="sm">
-                                            <Title
-                                                order={6}
-                                                c="gray.7"
-                                                size="sm"
-                                                fw={500}
-                                            >
-                                                Quick Templates
-                                            </Title>
-
-                                            <InstructionsTemplates
-                                                onSelect={(
-                                                    instruction: string,
-                                                ) => {
-                                                    form.setFieldValue(
-                                                        'instruction',
-                                                        form.values.instruction
-                                                            ? `${form.values.instruction}\n\n${instruction}`
-                                                            : instruction,
-                                                    );
-                                                }}
-                                            />
-                                        </Stack>
-                                        <Stack gap="sm">
-                                            <Box>
-                                                <Title
-                                                    order={6}
-                                                    c="gray.7"
-                                                    size="sm"
-                                                    fw={500}
-                                                >
-                                                    Guidelines
-                                                </Title>
-                                                <Text c="dimmed" size="xs">
-                                                    When writing instructions,
-                                                    consider the following
-                                                    guidelines to help the agent
-                                                    perform its tasks
-                                                    effectively.
-                                                </Text>
-                                            </Box>
-                                            <InstructionsGuidelines />
-                                            <Text c="dimmed" size="xs">
-                                                Visit our{' '}
-                                                <Anchor
-                                                    href="https://docs.lightdash.com/guides/ai-agents#writing-effective-instructions"
-                                                    target="_blank"
-                                                >
-                                                    docs
-                                                </Anchor>{' '}
-                                                to learn more about instructions
-                                                and how they work.
-                                            </Text>
-                                        </Stack>
-                                    </Stack>
-                                </Paper>
-
-                                <Paper p="xl">
-                                    <Group align="center" gap="xs" mb="md">
-                                        <Paper p="xxs" withBorder radius="sm">
-                                            <MantineIcon
-                                                icon={IconPlug}
-                                                size="md"
-                                            />
-                                        </Paper>
-                                        <Title order={5} c="gray.9" fw={700}>
-                                            Integrations
-                                        </Title>
-                                    </Group>
-                                    <Stack gap="sm">
-                                        <Group
-                                            align="center"
-                                            justify="space-between"
-                                            gap="xs"
-                                        >
-                                            <Title order={6}>Slack</Title>
-                                            <Group
-                                                c={
-                                                    slackChannelsConfigured
-                                                        ? 'green.04'
-                                                        : 'dimmed'
-                                                }
-                                                gap="xxs"
-                                                align="flex-start"
-                                            >
-                                                <MantineIcon
-                                                    icon={IconPointFilled}
-                                                    size={16}
-                                                />
-                                                <Text size="xs">
-                                                    {!slackInstallation?.organizationUuid
-                                                        ? 'Disabled'
-                                                        : !slackChannelsConfigured
-                                                        ? 'Channels not configured'
-                                                        : 'Enabled'}
-                                                </Text>
-                                            </Group>
-                                        </Group>
-
-                                        <LoadingOverlay
-                                            visible={isLoadingSlackInstallation}
-                                        />
-                                        {!slackInstallation?.organizationUuid ? (
-                                            <Paper
-                                                withBorder
-                                                p="sm"
-                                                style={{
-                                                    borderStyle: 'dashed',
-                                                    backgroundColor:
-                                                        'transparent',
-                                                }}
-                                            >
-                                                <Text
-                                                    size="xs"
-                                                    c="dimmed"
-                                                    ta="center"
-                                                >
-                                                    To enable AI agent
-                                                    interactions through Slack,
-                                                    please connect your Slack
-                                                    workspace in the{' '}
-                                                    <Anchor
-                                                        c="dimmed"
-                                                        underline="always"
-                                                        href="/generalSettings/integrations"
-                                                        target="_blank"
-                                                    >
-                                                        Integrations settings
-                                                    </Anchor>
-                                                    . Once connected, you can
-                                                    select channels where this
-                                                    agent will be available.
-                                                </Text>
-                                            </Paper>
-                                        ) : (
-                                            <Box>
-                                                <Stack gap="xs">
-                                                    <MultiSelect
-                                                        variant="subtle"
-                                                        readOnly={
-                                                            isLoadingSlackChannels ||
-                                                            isRefreshing
-                                                        }
-                                                        description={
-                                                            <>
-                                                                Select the
-                                                                channels where
-                                                                this agent will
-                                                                be available.
-                                                                {slackChannelsConfigured && (
-                                                                    <>
-                                                                        {' '}
-                                                                        Tag the
-                                                                        Slack
-                                                                        app{' '}
-                                                                        <Code>
-                                                                            @
-                                                                            {
-                                                                                slackInstallation.appName
-                                                                            }
-                                                                        </Code>{' '}
-                                                                        to get
-                                                                        started.
-                                                                    </>
-                                                                )}
-                                                            </>
-                                                        }
-                                                        labelProps={{
-                                                            style: {
-                                                                width: '100%',
-                                                            },
-                                                        }}
-                                                        label={'Channels'}
-                                                        limit={30}
-                                                        placeholder={
-                                                            isLoadingSlackChannels ||
-                                                            isRefreshing
-                                                                ? 'Loading channels, this might take a while if you have a lot of channels in your workspace'
-                                                                : 'Search channel(s)'
-                                                        }
-                                                        data={
-                                                            slackChannelOptions
-                                                        }
-                                                        value={form.values.integrations.map(
-                                                            (i) => i.channelId,
-                                                        )}
-                                                        searchable
-                                                        rightSectionPointerEvents="all"
-                                                        rightSection={
-                                                            isLoadingSlackChannels ||
-                                                            isRefreshing ? (
-                                                                <Loader size="xs" />
-                                                            ) : (
-                                                                <Tooltip
-                                                                    withArrow
-                                                                    withinPortal
-                                                                    label="Refresh Slack Channels"
-                                                                >
-                                                                    <ActionIcon
-                                                                        variant="transparent"
-                                                                        onClick={
-                                                                            refreshChannels
-                                                                        }
-                                                                    >
-                                                                        <MantineIcon
-                                                                            icon={
-                                                                                IconRefresh
-                                                                            }
-                                                                        />
-                                                                    </ActionIcon>
-                                                                </Tooltip>
-                                                            )
-                                                        }
-                                                        onChange={(value) => {
-                                                            form.setFieldValue(
-                                                                'integrations',
-                                                                value.map(
-                                                                    (v) =>
-                                                                        ({
-                                                                            type: 'slack',
-                                                                            channelId:
-                                                                                v,
-                                                                        } as const),
-                                                                ),
-                                                            );
-                                                        }}
-                                                    />
-                                                </Stack>
-                                            </Box>
-                                        )}
-                                    </Stack>
-                                </Paper>
-
-                                {!isCreateMode && (
-                                    <Paper p="xl" withBorder>
-                                        <Group align="center" gap="xs" mb="md">
-                                            <Paper
-                                                p="xxs"
-                                                withBorder
-                                                radius="sm"
-                                            >
-                                                <MantineIcon
-                                                    icon={IconAlertTriangle}
-                                                    size="md"
-                                                />
-                                            </Paper>
-                                            <Title
-                                                order={5}
-                                                c="gray.9"
-                                                fw={700}
-                                            >
-                                                Danger zone
-                                            </Title>
-                                        </Group>
-                                        <Group
-                                            gap="xs"
-                                            align="center"
-                                            justify="space-between"
-                                        >
-                                            <Box>
-                                                <Title
-                                                    order={6}
-                                                    c="gray.7"
-                                                    size="sm"
-                                                    fw={500}
-                                                >
-                                                    Delete agent
-                                                </Title>
-                                                <Text c="dimmed" size="xs">
-                                                    Deleting an agent will
-                                                    remove all its data and
-                                                    conversations.
-                                                </Text>
-                                            </Box>
-                                            <Button
-                                                variant="outline"
-                                                color="red"
-                                                onClick={handleDeleteClick}
-                                                leftSection={
-                                                    <MantineIcon
-                                                        icon={IconTrash}
-                                                    />
-                                                }
-                                            >
-                                                Delete
-                                            </Button>
-                                        </Group>
-                                    </Paper>
-                                )}
-                            </Stack>
-                        </form>
-                    </Tabs.Panel>
-
-                    <Tabs.Panel value="evals" pt="lg">
-                        <EvalTabLayout>
+                    {activeTab === 'evals' && (
+                        <EvalSectionLayout>
                             {runUuid ? (
                                 <EvalRunDetails
                                     projectUuid={projectUuid!}
@@ -1096,42 +343,16 @@ const ProjectAiAgentEditPage: FC<Props> = ({ isCreateMode = false }) => {
                                     evalUuid={evalUuid}
                                 />
                             ) : (
-                                <EvalsTab
+                                <EvalsSetup
                                     projectUuid={projectUuid!}
                                     agentUuid={actualAgentUuid!}
                                 />
                             )}
-                        </EvalTabLayout>
-                    </Tabs.Panel>
-                </Tabs>
-                <MantineModal
-                    opened={deleteModalOpen}
-                    onClose={handleCancelDelete}
-                    title="Delete Agent"
-                    icon={IconTrash}
-                    actions={
-                        <Group>
-                            <Button
-                                variant="subtle"
-                                onClick={handleCancelDelete}
-                            >
-                                Cancel
-                            </Button>
-                            <Button color="red" onClick={handleDelete}>
-                                Delete
-                            </Button>
-                        </Group>
-                    }
-                >
-                    <Stack gap="md">
-                        <Text>
-                            Are you sure you want to delete this agent? This
-                            action cannot be undone.
-                        </Text>
-                    </Stack>
-                </MantineModal>
-            </Stack>
-        </AiAgentEditPageLayout>
+                        </EvalSectionLayout>
+                    )}
+                </Box>
+            </AppShell.Main>
+        </AppShell>
     );
 };
